add security fix

+++ controller/userController.js
@@ -7,7 +7,7 @@ module.exports.profile = function(req, res){
         return res.redirect('/'); 
     } 
     //Fetching the al the todo's  data and passing it to the view i.e ejs template. 
-    Todo.find({user: req.user._id}, (err, document)=>{
+    Todo.find({user: req.user._id}, (err, documents)=>{
         if(err){
             console.log("Error while fetching documents!"); 
             return; 
@@ -16,16 +16,20 @@ module.exports.profile = function(req, res){
 
         //Modifying date and then passing it to ejs template. 
         // This will modify all dates {from:"2019-10-30 00:00:00.000Z" => TO: "Wed Oct 30 2019"}.
-        for(let todo of document){
+        for(let todo of documents){
             let date = new Date(todo.deadline); 
             dates.push(date.toDateString()); 
         }
 
-        res.render('user', {todos: document, date: dates, user: req.user}); 
+        res.render('user', {todos: documents, date: dates, user: req.user}); 
     }); 
 }; 
 
+
+//Logout Controller.
 module.exports.logout = function(req, res){
+    //If the user is authenticated then clear cookie. 
+    //  redirect to /index         
     if(req.isAuthenticated()){
         req.logout(); 
     }
@@ -35,6 +39,9 @@ module.exports.logout = function(req, res){
 
 //Controller for add-todo route i.e /add-todo.
 module.exports.addTodo = function(req, res){
+
+    //If the user is authenticated then only let the user to add the todo. 
+    //Else redirect it back to /index
     if(req.isAuthenticated()){
         var desc = req.body.description; 
         var date = req.body.date; 
@@ -72,18 +79,22 @@ module.exports.deleteTodo = function(req, res){
     Request that is comming looks like http://domain.com/user/delete-todo/5e953841ff05b644a8d1b80e,5e953841ff05b644a8d1b80e 
     this gets converted to an array of [5e953841ff05b644a8d1b80e,5e953841ff05b644a8d1b80e ](If 2 todo are selected similarly for n todo's)
     */
+
+    //Check if the user is authenticated or not and if not then don't give the permission to user to delete the todo. 
     if(req.isAuthenticated()){
         let arr = req.params.id.split(','); 
         arr.forEach(element => {
-            Todo.findByIdAndDelete(element, (err)=>{
-                if(err){
-                    console.log("Error while deleting document."); 
-                    return; 
-                }
-            })
+            if(element.user == req.user.id){
+                Todo.findByIdAndDelete(element, (err)=>{
+                    if(err){
+                        console.log("Error while deleting document."); 
+                        return; 
+                    }
+                });
+            }
         });
         return res.redirect('back'); 
     }else{
-        return res.redirect('/'); 
+        return res.redirect('back'); 
     }
 }
